+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 03/12/2020
+
+- Add new field `geostoreDataApi` to store geostore IDs for the Data API (alongside RW API geostore IDs).
+
 ## 19/11/2020
 
 - Add support for sorting on getting areas v2